wake-lock/wakelock-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that wake-lock is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://www.w3.org/TR/wake-lock/#dfn-wake-lock-feature">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('wake-lock', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise wake-lock.');
</script>
